requirements.txt
Flask>=2.0
Flask-WTF>=1.0
Flask-Login>=0.6
Flask-SQLAlchemy>=3.0
WTForms>=3.0
Werkzeug>=2.0
email-validator>=1.3
Pillow>=9.0
